Split stack profile registry modules
Move typed stack models, default profile data, template mappings, and registry selection logic into focused modules while keeping src.stack.profile compatibility exports.

Evidence: before refactor tests passed with 214 passed, ruff passed, and mypy passed. After refactor tests passed with 214 passed, ruff passed, and mypy passed. Generated output for TypeScript service, Rust Cloudflare Worker service, and hybrid monorepo matched baseline with empty diff -ru output. Tests were not modified.
