+++ src/stack/__init__.py
@@ -1,5 +1,5 @@
 """Stack profile registry for Kickstart scaffold options."""
 
-from src.stack.profile import stack_registry
+from src.stack.registry import stack_registry
 
 __all__ = ["stack_registry"]
